Add truncation for precision in fractional seconds #8

+++ include/another_toml/types.hpp
@@ -41,7 +41,7 @@ namespace another_toml
 		std::uint8_t hours = {},
 			minutes = {},
 			seconds = {};
-		float seconds_frac = {};
+		double seconds_frac = {};
 	};
 
 	// Compound date/time type
